prefer std::u?int[0-9]+_t and include <cstdint> else include <stdint.h> (#2302)

+++ include/igl/copyleft/cgal/peel_outer_hull_layers.h
@@ -30,7 +30,7 @@ namespace igl
         typename DerivedF,
         typename DerivedI,
         typename Derivedflip>
-      IGL_INLINE size_t peel_outer_hull_layers(
+      IGL_INLINE int peel_outer_hull_layers(
         const Eigen::PlainObjectBase<DerivedV > & V,
         const Eigen::PlainObjectBase<DerivedF > & F,
         Eigen::PlainObjectBase<DerivedI > & I,
